Major code fix 2
This fixes attempting to sort a pocket multiple times that is empty, or only has one item, eventually leading to a crash.

I've also added scroll arrows to now draw when shifting an item up or down in the bag, purely for cosmetic reasons.

Finally, a tiny bit of text adjustment for better logical flow.