Code changes, flash fire revert
Reverted FLASH firing automatically, however you can ignore that if you like it and wish to keep it.

Changed party menu functionality - TELEPORT and FLY are only visible to use in places that allow it, and as such it won't appear in the party menu inside houses or buildings, for example. FLASH is only visible to use inside of caves and only if you haven't lit up the cave already. CUT is only visible to use if you're in front of and *facing* a cuttable square of grass, in front of a cuttable tree, or in front of the dotted hole entrance door. DIG is only visible to use in places that allow for using an escape rope. MILK DRINK and SOFT-BOILED are now only visible to use if your current HP is >=20% of your max HP, but I couldn't figure out how to hide their use if all other party members are at full health. SWEET SCENT is now only visible to use if you're in places that allow for flying/teleporting OR using an escape rope. Still testing these, more changes/reversions to come?