Code and formatting changes
The location where you last saved is now printed on the continue box on the main menu. I had to shift things around to make it fit. The save screen in game has also been shifted around, and the player name has been changed from the awful bulky font to the one the rest of the text in that box uses.

Possibly has bugs, will test and revisit at a later time.