Pre evolution move relearning
Credit to devolov

"Goal: Make it so reteaching a move through the move relearner allows for moves that the Pre-Evoltuions know to still be taught.

For example, if you have a level 60 Clyster and want to reteach it Ice Beam, you cannot in the normal game. That's because Shellder learns Ice Beam at level 49, but Cloyster never learns it. This will make it so the moves of all Pre-evoluions are in the move tutor's menu."