Daycare cap fix 2
Another fix for this annoying feature :(

If you put a pokemon *over* the level cap in the daycare, it would get rounded down on return and the daycare man text would be garbled.
Now, pokemon at or over the level cap are treated the same as eggs, and categorically rejected from being stored.