Rock Tunnel changes
Block the registered item menu from being able to activate in ROCK TUNNEL unless FLASH is active. The item select box would originally draw under the darkness of the cave, and this stops that from happening, as it's borderline impossible to notice the box is active if you pressed the button by mistake.