Read comment
This branch is a fucking mess, and will be getting redone in the near future. Bug fixes and some minor code refactoring have been moved to their own combined branch and will be the defacto branch for anything of that nature. I will be making RTC its own branch, and RSification its own branch, to make things easier on myself to maintain and update.