Context wrap
Context menus now wrap around instead of being forced to scroll all the way to the bottom or back up to the top. Only minor testing has been done, could very well have bugs.

If the scroll selection goes outside of the box, then you probably need to edit sContextMenuNumOptions for whatever that box is controlling. An example for this would be PrintNameChoiceOptions.