Evolve manually
Credit to Lunos

"This feature branch adds a new option to the game's party screen through which you can evolve a Pokémon if their evolution conditions are met.

This naturally excludes Pokémon that evolve through the usage of items or that evolve by trading."

A bit of minor changing from the source to fix a couple of FRLG specific issues. Tested working with no obvious problems.